+++ CHANGELOG.md
@@ -9,6 +9,7 @@ All notable changes to this project will be documented in this file.
 ### Changed
 
 ### Fixed
+- Fixed BlockNote drag-handle delete getting the side menu stuck at a stale position after removing a block.
 
 ## [0.1.0] - 2026-03-10